+++ changelog.md
@@ -10,3 +10,12 @@ All notable changes to BaseMobileTasks-ane will be documented in this file.
 
 #### Changed
 - Removed extension id from ane.properties
+
+### 1.2.0
+#### Added
+- Internet connection check
+
+#### Changed
+- Renamed Context.java to MobileTasksContext.java
+- Now catching Throwables
+- General clean up